tx: split out tx sections into dedicated mod
